StartAppSDK Release 4.5.0

+++ StartApp.framework/Versions/A/Headers/STABannerView.h
@@ -4,20 +4,22 @@
 //
 //  Created by StartApp on 11/13/13.
 //  Copyright (c) 2013 StartApp. All rights reserved.
-//  SDK version 4.4.3
+//  SDK version 4.5.0
 
 #import <UIKit/UIKit.h>
 #import "STABannerSize.h"
 #import "STAAdPreferences.h"
 
 @class STABannerView;   // Forward decleration
+
 @protocol STABannerDelegateProtocol <NSObject>
+
 @optional
-- (void) bannerAdIsReadyToDisplay:(STABannerView*)banner;
-- (void) didDisplayBannerAd:(STABannerView*)banner;
-- (void) failedLoadBannerAd:(STABannerView*)banner withError:(NSError *)error;
-- (void) didClickBannerAd:(STABannerView*)banner;
-- (void) didCloseBannerInAppStore:(STABannerView*)banner;
+- (void) bannerAdIsReadyToDisplay:(STABannerView *)banner;
+- (void) didDisplayBannerAd:(STABannerView *)banner;
+- (void) failedLoadBannerAd:(STABannerView *)banner withError:(NSError *)error;
+- (void) didClickBannerAd:(STABannerView *)banner;
+- (void) didCloseBannerInAppStore:(STABannerView *)banner;
 
 @end
 
@@ -33,18 +35,17 @@ typedef enum {
 - (id)initWithSize:(STABannerSize)size autoOrigin:(STAAdOrigin)origin withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate;
 - (id)initWithSize:(STABannerSize)size autoOrigin:(STAAdOrigin)origin adPreferences:(STAAdPreferences *)adPreferences withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate;
 
-- (id)initWithSize:(STABannerSize)size origin:(CGPoint)origin withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate withAdTag:(NSString*)adTag;
-- (id)initWithSize:(STABannerSize)size origin:(CGPoint)origin adPreferences:(STAAdPreferences *)adPreferences withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate withAdTag:(NSString*)adTag;
-- (id)initWithSize:(STABannerSize)size autoOrigin:(STAAdOrigin)origin withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate withAdTag:(NSString*)adTag;
-- (id)initWithSize:(STABannerSize)size autoOrigin:(STAAdOrigin)origin adPreferences:(STAAdPreferences *)adPreferences withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate withAdTag:(NSString*)adTag;
-
+- (id)initWithSize:(STABannerSize)size origin:(CGPoint)origin withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate withAdTag:(NSString *)adTag;
+- (id)initWithSize:(STABannerSize)size origin:(CGPoint)origin adPreferences:(STAAdPreferences *)adPreferences withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate withAdTag:(NSString *)adTag;
+- (id)initWithSize:(STABannerSize)size autoOrigin:(STAAdOrigin)origin withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate withAdTag:(NSString *)adTag;
+- (id)initWithSize:(STABannerSize)size autoOrigin:(STAAdOrigin)origin adPreferences:(STAAdPreferences *)adPreferences withDelegate:(id<STABannerDelegateProtocol>)bannerDelegate withAdTag:(NSString *)adTag;
 
 - (void)loadAd;
 
-- (void)setSTABannerAdTag:(NSString *) adTag;
+- (void)setSTABannerAdTag:(NSString *)adTag;
 
-- (void)setSTABannerSize:(STABannerSize) size;
-- (void)setOrigin:(CGPoint) origin;
+- (void)setSTABannerSize:(STABannerSize)size;
+- (void)setOrigin:(CGPoint)origin;
 - (void)setSTAAutoOrigin:(STAAdOrigin)origin;
 
 - (void)setAdPreferneces:(STAAdPreferences *)adPreferences;
@@ -53,10 +54,10 @@ typedef enum {
 - (void)showBanner;
 - (BOOL)isVisible;
 
--(void)addSTABannerToCell:(UITableViewCell *)cell withIndexPath:(NSIndexPath *)indexPath atIntexPathSection:(int)section repeatEach:(int)each;
--(void)addSTABannerToCell:(UITableViewCell *)cell withIndexPath:(NSIndexPath *)indexPath atIntexPathSection:(int)section;
+-(void)addSTABannerToCell:(UITableViewCell *)cell withIndexPath:(NSIndexPath *)indexPath atIntexPathSection:(NSInteger)section repeatEach:(NSInteger)each;
+-(void)addSTABannerToCell:(UITableViewCell *)cell withIndexPath:(NSIndexPath *)indexPath atIntexPathSection:(NSInteger)section;
 
--(void)addSTABannerToCell:(UITableViewCell *)cell withIndexPath:(NSIndexPath *)indexPath atIntexPathRow:(int)row repeatEach:(int)each;
--(void)addSTABannerToCell:(UITableViewCell *)cell withIndexPath:(NSIndexPath *)indexPath atIntexPathRow:(int)row;
+-(void)addSTABannerToCell:(UITableViewCell *)cell withIndexPath:(NSIndexPath *)indexPath atIntexPathRow:(NSInteger)row repeatEach:(NSInteger)each;
+-(void)addSTABannerToCell:(UITableViewCell *)cell withIndexPath:(NSIndexPath *)indexPath atIntexPathRow:(NSInteger)row;
 
 @end
